fixed the delete btn

+++ public/js/deletemovie.js
@@ -1,15 +1,19 @@
-const movieId = document.querySelector('input[name="movie-id"]').value;
 
-const deleteHandler = async () => {
-  const delete_movie = await fetch(`/api/movies/${movieId}`, {
-    method: "DELETE",
-  });
+const deleteHandler = async (event) => {
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
+    const delete_movie = await fetch(`/api/movies/${id}`, {
+      method: "DELETE",
+    });
 
-  if (delete_movie.ok) {
-    document.location.replace("/dashboard");
-  } else {
-    alert("Failed to delete movie");
+    if (delete_movie.ok) {
+      document.location.replace("/dashboard");
+    } else {
+      alert("Failed to delete movie");
+    }
   }
 };
 
-document.querySelector("#deleteBtn").addEventListener("click", deleteHandler);
+[...document.querySelectorAll(".deleteBtn")].forEach(function (item) {
+  item.addEventListener("click", deleteHandler);
+});
